requirements.txt
langchain
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langchain-ollama
watchdog
langserve
fastapi
uvicorn
sse_starlette
pydantic
bs4
pypdf
chromadb
faiss-cpu
lancedb
arxiv
wikipedia
langchainhub
groq
langchain-groq